Allows to append entities to a resource
Currently the only ways to change the entities of a resource is to
replace all the entities with PATCH call on the resource.

This change allows to append new entities to a resource without touching
entities already associated with it.

+++ gnocchi/indexer/sqlalchemy.py
@@ -239,6 +239,7 @@ class SQLAlchemyIndexer(indexer.IndexerDriver):
 
     def update_resource(self, resource_type,
                         uuid, ended_at=_marker, entities=_marker,
+                        append_entities=False,
                         **kwargs):
         resource_cls = self._resource_type_to_class(resource_type)
         session = self.engine_facade.get_session()
@@ -271,8 +272,9 @@ class SQLAlchemyIndexer(indexer.IndexerDriver):
                             r.type, attribute)
 
             if entities is not _marker:
-                session.query(ResourceEntity).filter(
-                    ResourceEntity.resource_id == uuid).delete()
+                if not append_entities:
+                    session.query(ResourceEntity).filter(
+                        ResourceEntity.resource_id == uuid).delete()
                 for name, e in entities.iteritems():
                     session.add(ResourceEntity(resource_id=uuid,
                                                entity_id=e,
